Fix editor API
