+++ package.json
@@ -52,7 +52,7 @@
     "clean-webpack-plugin": "^3.0.0",
     "eslint": "8.19.0",
     "eslint-config-airbnb-typescript-prettier": "5.0.0",
-    "eslint-plugin-jest": "26.5.3",
+    "eslint-plugin-jest": "^26.5.3",
     "husky": "^4.2.5",
     "jest": "28.1.1",
     "lint-staged": "^10.2.13",
