+++ tsconfig.json
@@ -6,7 +6,7 @@
     "module": "esnext", // specify module code generation
     "jsx": "react", // use typescript to transpile jsx to js
     "target": "es6", // specify ECMAScript target version
-    "lib": ["es2020", "dom"],
+    "lib": ["es2021", "dom"],
     "allowJs": true, // allow a partial TypeScript and JavaScript codebase
     "experimentalDecorators": true,
     "moduleResolution": "node",
